feat: add EmailService for sending emails using PHPMailer
- Implemented EmailService class to handle email sending functionality.
- Configured SMTP settings for Mailtrap.
- Added method to send general emails and a specific method for notifying about newly created dishes.

+++ src/routes/web.php
@@ -194,6 +194,12 @@ function handleRoute($uri) {
         return;
     }
     
+    if (preg_match('/^\/admin\/usuarios\/cambiar-rol\/(\d+)$/', $uri, $matches)) {
+        $controller = new AdminController();
+        $controller->cambiarRolUsuario($matches[1]);
+        return;
+    }
+    
     if (preg_match('/^\/admin\/usuarios\/eliminar\/(\d+)$/', $uri, $matches)) {
         $controller = new AdminController();
         $controller->eliminarUsuario($matches[1]);
